refactor: move run_code_graph_pipeline to cognify/code_graph_pipeline
